scenarios/rag/custom-rag-app/requirements.txt
azure-ai-projects
azure-ai-inference
azure-ai-identity
azure-search-documents
pandas
python-dotenv
open-telemetry-api
